backend/requirements.txt
fastapi
uvicorn
redis
python-dotenv
boto3
langchain
llama-index
elasticsearch
unstructured
python-multipart
pydantic
redis-om
python-jose
langchain-community
passlib[bcrypt]
unstructured[pdf]
langchain-google-genai
google-generativeai
faiss-cpu


